backend/requirements.txt
antlr4-python3-runtime==4.9.3
asgiref==3.8.1
contourpy==1.3.0
cycler==0.12.1
Django==5.1.1
django-environ==0.11.2
djangorestframework==3.15.2
filelock==3.16.1
fonttools==4.54.0
fsspec==2024.9.0
hydra-core==1.3.2
imageio==2.35.1
iopath==0.1.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
portalocker==2.10.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
-e git+https://github.com/facebookresearch/segment-anything-2.git@7e1596c0b6462eb1d1ba7e1492430fed95023598#egg=SAM_2
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
